requirements.txt
google-cloud-texttospeech
flask
python-vlc